refactor(backend): use decimal for money amounts

+++ src/backend/formatting.py
@@ -1,7 +1,9 @@
 """German locale formatters for prices and quantities."""
 
+from decimal import Decimal
 
-def format_quantity(value: float) -> str:
+
+def format_quantity(value: Decimal) -> str:
     """Format a quantity value with one decimal,
     German locale (e.g. 1.5 -> 1,5; 2.0 -> 2)."""
     if value % 1 == 0:
@@ -9,7 +11,7 @@ def format_quantity(value: float) -> str:
     return f"{value:.1f}".replace(".", ",")
 
 
-def format_price(value: float) -> str:
+def format_price(value: Decimal) -> str:
     """Format a price value with thousands separator and two decimals,
     German locale (e.g. 1234.5 -> 1.234,50€)."""
     return f"{value:,.2f}".replace(",", "X").replace(".", ",").replace("X", ".") + "€"
